+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 
-- Update: Set start of the week to Monday for the date fields
+- New: Added an option to hide time fields on the Flight Record page.
+- Update: Reorganized action buttons on the Flight Record page.
+- Update: Set start of the week to Monday for the date fields.
 - New: Added badges on the navigation panel for very soon expiring and expired licenses, configurable from the Settings page.
 
 ## [3.14.0] - 07.10.2025